Handle typed Tesla fields
Starting in August, Tesla will always behave as if prefer_typed is set
to true. Here we begin using the proper types, though we can't assume
that fields will always be typed, because of older firmware.

There is a decent amount of reorganization here, and I think the
generate code is better.