Handle typed Tesla fields (#170)
* Handle typed Tesla fields

Starting in August, Tesla will always behave as if prefer_typed is set
to true. Here we begin using the proper types, though we can't assume
that fields will always be typed, because of older firmware.

There is a decent amount of reorganization here, and I think the
generate code is better.

* Fix commented rule, and trim trailing

* It's JSON

* Package comment

* Get rid of the "outer functions"

Also make some comments in the template. Tests are passing.

* Try to vain to standardize on tabs  in the template

* More tab stuff, another comment

* Bring back the unit comments

* Try to shave some lines

* Add tests for the typed case

* Add more comments, skip generation if no targets

* Try to get this function size down again

* Remove this TeslaTypeName field that we didn't use

* More comments. Down with agile

* Test errors'

* More careful with window states

* Update VSS CSV so that we can auto-convert range

* Removed now-unnecessary manual range conversion

* Correct the Tesla type of ChargeLimitSoc to int32

* Get rid of these needless Fatalfs

* Add crummy README

* Finish sentence